repo() only returns Git repositories
Some repositories on my path have subdirectories with
the same name as a repository. For example:

    $HOME/code/.../repo1/docs/repo2
    $HOME/code/.../repo2/

which means that repo() will sometimes cd into the subdirectory
of a repository which is unrelated to the repository passed
in on the command line.

This commit ensures that repo /only/ cds into Git repositories.

This will need changing again if I start to use another VCS.