Scissors to delimit the end of a Git commit msg
Without this option and line starting with a hash
will be ignored by Git when writing a commit message.
That means that if you have code in your commit
message that uses hashes to indicate comments, those
lines will not be included in the commit message.

With this option, everything underneath a 'scissors'
line will be removed from the message instead.

Scissors lines look like ' >8 ' surrounded by
dashes and prefixed by a hash.