Repo command accepts - as an argument
Previously 'repo -' failed becauase the command could
not find a directory that was a Git repository called '-'.

This commit changes that behaviour, so that 'repo -'
just runs 'cd' to the last directory that was used as
the CWD. Note that this will allow 'repo' to cd to
a directory which was not a Git repository.